refactor: split identify_opponent handlers into a mixin subpackage
Convert widgets/frames/identify_opponent/handlers.py into a handlers/
subpackage, one mixin per subsystem, composed back into the single
MTGOpponentDeckSpyHandlersMixin so the public import path
(widgets.frames.identify_opponent.handlers) is unchanged.

- calculator.py -> CalculatorMixin
- radar.py -> RadarMixin
- polling.py -> OpponentPollingMixin
- archetype.py -> ManualArchetypeMixin
- persistence.py -> PersistenceMixin
- window_placement.py -> WindowPlacementMixin
- lifecycle.py -> LifecycleMixin

Promote the inline typed-attribute block into protocol.py
(MTGOpponentDeckSpyProto); each mixin types self via the
TYPE_CHECKING `_Base = Proto else object` idiom used elsewhere
in the app. Pure structural move; no behavior change.
