+++ widgets/frames/identify_opponent/protocol.py
@@ -1,0 +1,102 @@
+"""Shared ``self`` contract that the opponent-tracker handler mixins assume.
+
+Promotes the inline typed-attribute block that previously lived at the top of
+``MTGOpponentDeckSpyHandlersMixin`` into a single :class:`Protocol`, so each
+handler mixin in :mod:`.handlers` can type its ``self`` consistently via the
+``TYPE_CHECKING: _Base = Proto else object`` idiom used elsewhere in the app.
+"""
+
+from __future__ import annotations
+
+import threading
+from typing import TYPE_CHECKING, Any, Protocol
+
+import wx
+
+from widgets.panels.compact_radar_panel import CompactRadarPanel
+from widgets.panels.compact_sideboard_panel import CompactSideboardPanel
+
+if TYPE_CHECKING:
+    from services.metagame_service import MetagameService
+    from services.radar_service import RadarData, RadarService
+
+
+class MTGOpponentDeckSpyProto(Protocol):
+    """Cross-mixin ``self`` surface for the opponent-tracker frame."""
+
+    # --- typed attribute declarations for type-checker clarity ---
+    CACHE_TTL: int
+    POLL_INTERVAL_MS: int
+
+    cache: dict[str, dict[str, Any]]
+    player_name: str
+    last_seen_decks: dict[str, str]
+
+    _poll_timer: wx.Timer
+    _saved_position: list[int] | None
+    _bg_worker: Any
+    _poll_generation: int
+    _poll_in_progress: bool
+    _watching_enabled: bool
+    _manual_archetype_loaded: bool
+
+    radar_service: RadarService
+    metagame_service: MetagameService
+    current_radar: RadarData | None
+    _radar_worker_thread: threading.Thread | None
+    _radar_cancel_requested: bool
+    _last_radar_archetype: str
+    _last_guide_archetype: str
+
+    deck_label: wx.StaticText
+    status_label: wx.StaticText
+    load_arch_btn: wx.Button
+    calc_result_label: wx.StaticText
+    spin_deck_size: wx.SpinCtrl
+    spin_copies: wx.SpinCtrl
+    spin_drawn: wx.SpinCtrl
+    spin_target: wx.SpinCtrl
+    radar_panel: CompactRadarPanel
+    sideboard_panel: CompactSideboardPanel
+
+    # Provided by the frame / properties mixin and consumed by the handlers.
+    _locale: str | None
+    controller: Any
+
+    def _t(self, key: str, **kwargs: object) -> str: ...
+
+    # Cross-mixin handler methods.
+    def _unload_manual_archetype(self) -> None: ...
+    def _load_archetype_manually(self, fmt: str, archetype: str) -> None: ...
+    def _update_guide_display(self) -> None: ...
+    def _on_calculate(self, _event: wx.CommandEvent | None) -> None: ...
+    def _trigger_radar_load(self) -> None: ...
+    def _generate_radar_worker(self, archetype_dict: dict[str, Any], format_name: str) -> None: ...
+    def _display_radar(self, radar: RadarData) -> None: ...
+    def _clear_radar_display(self) -> None: ...
+    def _start_polling(self) -> None: ...
+    def _stop_watching(self) -> None: ...
+    def _submit_poll(self) -> None: ...
+    def _poll_worker(self, generation: int, current_player: str) -> dict: ...
+    def _on_poll_error(self, exc: Exception) -> None: ...
+    def _apply_poll_result(self, result: dict) -> None: ...
+    def _lookup_decks_all_formats(
+        self, opponent_name: str, *, force: bool = False
+    ) -> dict[str, str]: ...
+    def _refresh_opponent_display(self) -> None: ...
+    def _save_config(self) -> None: ...
+    def _save_cache(self) -> None: ...
+    def _load_cache(self) -> None: ...
+    def _load_config(self) -> None: ...
+    def _place_beside_parent(self) -> None: ...
+
+    # wx.Frame surface used by the handlers.
+    def GetPosition(self) -> wx.Point: ...
+    def GetParent(self) -> wx.Window | None: ...
+    def GetRect(self) -> wx.Rect: ...
+    def GetSize(self) -> wx.Size: ...
+    def SetSize(self, *args: Any) -> None: ...
+    def SetPosition(self, pt: wx.Point) -> None: ...
+    def SetMinSize(self, size: wx.Size) -> None: ...
+    def SetBackgroundColour(self, colour: Any) -> bool: ...
+    def IsShown(self) -> bool: ...
